fix(jni): fix boxed value unboxing and volatile increment issues
Fix BX_UNBOXING_IMMEDIATELY_REBOXED in JniGlobal.convertToWasmValue()
by passing Long objects directly to funcref/externref methods instead
of unboxing to long primitive and reboxing.

Fix VO_VOLATILE_INCREMENT issues by converting volatile int fields to
AtomicInteger:
- CompilationCache.CacheEntry.accessCount
- OptimizedMarshalling.MarshallingPlan.useCount

This ensures thread-safe increments and resolves 3 SpotBugs warnings
(1 medium, 2 medium).

+++ wasmtime4j-jni/src/main/java/ai/tegmentum/wasmtime4j/jni/JniGlobal.java
@@ -401,7 +401,7 @@ public final class JniGlobal extends JniResource implements WasmGlobal {
           return WasmValue.funcref(null);
         }
         if (value instanceof Long) {
-          return WasmValue.funcref(((Long) value).longValue());
+          return WasmValue.funcref(value);
         }
         throw new WasmTypeException("Expected Long or null for funcref, got " + value.getClass());
       case "externref":
@@ -409,7 +409,7 @@ public final class JniGlobal extends JniResource implements WasmGlobal {
           return WasmValue.externref(null);
         }
         if (value instanceof Long) {
-          return WasmValue.externref(((Long) value).longValue());
+          return WasmValue.externref(value);
         }
         throw new WasmTypeException("Expected Long or null for externref, got " + value.getClass());
       default:
